Refactor: modificar para melhoria do menu lateral.
