+++ Elogica-RH-Frontend/src/app/shared/components/menu-lateral/menu-lateral.component.html
@@ -6,7 +6,9 @@
       <div *ngFor="let menuPai of menusPais">
         <div class="d-flex align-items-center menu-pai" [class.hovered]="menusAbertos[menuPai.id]"
           style="width: 151px; height: 35px; padding: 8px; cursor: pointer;"
-          (click)="menusFilhos[menuPai.id] ? toggleMenu(menuPai.id) : null">
+          (click)="menusFilhos[menuPai.id] ? toggleMenu(menuPai.id) : null"
+
+          >
           <i [class]="menuPai.icone + ' me-2'"></i>
           <div>{{ menuPai.titulo }}</div>
 
@@ -17,8 +19,12 @@
 
         <!-- Menus Filhos -->
         <div *ngIf="menusFilhos[menuPai.id] && menusAbertos[menuPai.id]" class="d-flex flex-column ms-4">
-          <div *ngFor="let menuFilho of menusFilhos[menuPai.id]"
-            class="d-flex align-items-center w-75 px-2 py-1 menu-filho">
+          <div
+            *ngFor="let menuFilho of menusFilhos[menuPai.id]"
+            class="d-flex align-items-center w-75 px-2 py-1 menu-filho"
+            [routerLink]="menuFilho.url"
+            style="cursor: pointer;"
+            >
             {{ menuFilho.titulo }}
           </div>
         </div>
@@ -26,5 +32,6 @@
     </div>
   </aside>
   <main>
+    <router-outlet></router-outlet>
   </main>
 </div>
